requirements.txt
redis==3.5.3
celery==5.4.0
Flask==2.0.2
Flask-Cors==3.0.10
requests==2.32.3
Werkzeug==2.2.2
requests
flower
beautifulsoup4==4.8.1
gunicorn
python-dotenv
selenium==4.22.0
mypy
pydantic
